Fix interpretation and data processing error
I misread the documentation on the incart database as is each individual
header file was for one patient. That is not the case: all patients have
recorded multiple files (typically two files) so the total patient
number is 32.

Fixed the data processing in R by including a patient identifier based
on the recorded files per patient and then updated the analyses.